Make StaticHiveCluster injectable
