--- presto-hive/src/test/java/com/facebook/presto/hive/TestHivePluginConfig.java
@@ -1,47 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.facebook.presto.hive;
-
-import com.google.common.collect.ImmutableMap;
-import org.testng.annotations.Test;
-
-import java.net.URI;
-import java.util.Map;
-
-import static io.airlift.configuration.testing.ConfigAssertions.assertFullMapping;
-import static io.airlift.configuration.testing.ConfigAssertions.assertRecordedDefaults;
-import static io.airlift.configuration.testing.ConfigAssertions.recordDefaults;
-
-public class TestHivePluginConfig
-{
-    @Test
-    public void testDefaults()
-    {
-        assertRecordedDefaults(recordDefaults(HivePluginConfig.class)
-                .setMetastoreUri(null));
-    }
-
-    @Test
-    public void testExplicitPropertyMappings()
-    {
-        Map<String, String> properties = new ImmutableMap.Builder<String, String>()
-                .put("hive.metastore.uri", "thrift://localhost:9083")
-                .build();
-
-        HivePluginConfig expected = new HivePluginConfig()
-                .setMetastoreUri(URI.create("thrift://localhost:9083"));
-
-        assertFullMapping(properties, expected);
-    }
-}
